backend/api_handlers.py
"""
API Handlers for domain-specific external services
"""

import os
import json
import aiohttp
import asyncio
from typing import Dict, Any, Optional
from dotenv import load_dotenv
import google.generativeai as genai

# Load environment variables
load_dotenv()

# Configure Google Generative AI
GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY", "")
if GOOGLE_API_KEY:
    genai.configure(api_key=GOOGLE_API_KEY)


class DomainApiHandler:
    """Handles API calls to various domain-specific services."""
    
    def __init__(self):
        """Initialize the API handler."""
        self.session = None
        self.api_keys = {
            "bloomberg": os.getenv("BLOOMBERG_API_KEY", ""),
            "courtlistener": os.getenv("COURTLISTENER_API_KEY", ""),
            "education": os.getenv("EDUCATION_API_KEY", ""),
            "deepchem": os.getenv("DEEPCHEM_API_KEY", ""),
            "support": os.getenv("SUPPORT_API_KEY", ""),
        }
    
    async def _get_session(self):
        """Get or create an aiohttp session."""
        if self.session is None or self.session.closed:
            self.session = aiohttp.ClientSession()
        return self.session
    
    async def query_bloomberg_api(self, query: str) -> Dict[str, Any]:
        """Query the Bloomberg API for financial data.
        
        Args:
            query: The query string
            
        Returns:
            The API response data
        """
        try:
            # Simulate API call for now
            await asyncio.sleep(0.5)  # Simulate network delay
            
            # Mock response
            return {
                "summary": f"Financial analysis for: {query}",
                "market_data": {
                    "indices": {"S&P 500": 4200.5, "NASDAQ": 14300.2},
                    "currencies": {"USD/EUR": 0.85, "USD/JPY": 110.2}
                },
                "metrics": {
                    "response_time_ms": 120,
                    "data_freshness_sec": 60
                }
            }
        except Exception as e:
            return {"error": str(e), "summary": "Error retrieving financial data"}
    
    async def query_courtlistener_api(self, query: str) -> Dict[str, Any]:
        """Query the CourtListener API for legal information.
        
        Args:
            query: The query string
            
        Returns:
            The API response data
        """
        try:
            # Simulate API call for now
            await asyncio.sleep(0.5)  # Simulate network delay
            
            # Mock response
            return {
                "summary": f"Legal analysis for: {query}",
                "cases": [
                    {"title": "Example v. Test", "citation": "123 F.3d 456", "year": 2022},
                    {"title": "Sample v. Demo", "citation": "789 F.3d 012", "year": 2021}
                ],
                "metrics": {
                    "response_time_ms": 150,
                    "total_cases": 2
                }
            }
        except Exception as e:
            return {"error": str(e), "summary": "Error retrieving legal data"}
    
    async def query_education_api(self, query: str) -> Dict[str, Any]:
        """Query the Education API for learning content.
        
        Args:
            query: The query string
            
        Returns:
            The API response data
        """
        try:
            # Simulate API call for now
            await asyncio.sleep(0.5)  # Simulate network delay
            
            # Mock response
            return {
                "summary": f"Educational content for: {query}",
                "learning_modules": [
                    {"title": "Introduction", "difficulty": "beginner"},
                    {"title": "Advanced Concepts", "difficulty": "intermediate"}
                ],
                "metrics": {
                    "response_time_ms": 90,
                    "personalization_score": 0.85
                }
            }
        except Exception as e:
            return {"error": str(e), "summary": "Error retrieving educational content"}
    
    async def query_deepchem_api(self, query: str) -> Dict[str, Any]:
        """Query the DeepChem API for healthcare information.
        
        Args:
            query: The query string
            
        Returns:
            The API response data
        """
        try:
            # Simulate API call for now
            await asyncio.sleep(0.5)  # Simulate network delay
            
            # Mock response
            return {
                "summary": f"Healthcare analysis for: {query}",
                "medical_data": {
                    "recommendations": ["Recommendation 1", "Recommendation 2"],
                    "references": ["Journal of Medicine, 2023", "Healthcare Review, 2022"]
                },
                "metrics": {
                    "response_time_ms": 200,
                    "confidence_score": 0.92
                }
            }
        except Exception as e:
            return {"error": str(e), "summary": "Error retrieving healthcare data"}
    
    async def query_support_api(self, query: str) -> Dict[str, Any]:
        """Query the Customer Support API.
        
        Args:
            query: The query string
            
        Returns:
            The API response data
        """
        try:
            # Simulate API call for now
            await asyncio.sleep(0.5)  # Simulate network delay
            
            # Mock response
            return {
                "summary": f"Support response for: {query}",
                "resolution_steps": [
                    "Step 1: Verify account information",
                    "Step 2: Check system status",
                    "Step 3: Reset configuration"
                ],
                "metrics": {
                    "response_time_ms": 80,
                    "satisfaction_prediction": 0.88
                }
            }
        except Exception as e:
            return {"error": str(e), "summary": "Error retrieving support information"}
    
    async def query_gemini_api(self, query: str, domain: str) -> Dict[str, Any]:
        """Query the Google Gemini API.
        
        Args:
            query: The query string
            domain: The domain for context
            
        Returns:
            The API response data
        """
        try:
            if not GOOGLE_API_KEY:
                # Mock response if API key is not available
                await asyncio.sleep(0.5)  # Simulate network delay
                return {
                    "summary": f"{domain.capitalize()} analysis for: {query}",
                    "generated_content": f"This is a simulated response for {domain} query: {query}",
                    "metrics": {
                        "response_time_ms": 150,
                        "tokens_used": 120
                    }
                }
            
            # Use actual Gemini API
            model = genai.GenerativeModel('gemini-pro')
            
            # Add domain-specific context to the prompt
            domain_context = f"You are an expert in {domain}. "
            prompt = domain_context + query
            
            # Generate response
            response = model.generate_content(prompt)
            
            return {
                "summary": response.text[:100] + "..." if len(response.text) > 100 else response.text,
                "generated_content": response.text,
                "metrics": {
                    "response_time_ms": 200,
                    "model": "gemini-pro"
                }
            }
        except Exception as e:
            return {"error": str(e), "summary": f"Error processing {domain} query"}
    
    async def close(self):
        """Close the aiohttp session."""
        if self.session and not self.session.closed:
            await self.session.close()