backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
sqlalchemy==2.0.28
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sentence-transformers==2.5.1
torch>=2.0.0
transformers>=4.30.0
pathway==0.8.4
faiss-cpu>=1.7.4
websockets>=12.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
google-generativeai>=0.3.0
litellm>=1.35.0
pinecone-client>=2.2.1
numpy>=1.24.0
jinja2>=3.0.0 